ETL/extract/requirements.txt
requests==2.31.0
python-dateutil==2.9.0
APScheduler==3.10.4
pika==1.3.2
